Switching to guaranteed synchronization
